KVM: PPC: Book3S HV: add a VC_BASE control to the XIVE native device
The ESB MMIO region controls the interrupt sources of the guest. QEMU
will query an fd (GET_ESB_FD ioctl) and map this region at a specific
address for the guest to use and the guest will obtain this
information using the H_INT_GET_SOURCE_INFO hcall. To inform KVM of
the address setting used by QEMU, let's add a VC_BASE control to the
KVM XIVE device

Signed-off-by: Cédric Le Goater <clg@kaod.org>