KVM: PPC: Book3S HV: add passthrough support
Clear the ESB pages from the VMA of the IRQ being pass through to the
guest and let the fault handler repopulate the VMA when the ESB pages
are accessed for an EOI or for a trigger.

Storing the VMA under the KVM XIVE device is a little ugly. Needs more
thinking.

Signed-off-by: Cédric Le Goater <clg@kaod.org>