vfio/mdev: introduce mdev_bus_is_mdev() helper
The type1 IOMMU driver has some code to guess if the IOMMU group
is linked to a mediated device. This logic isn't type1 specific
though, so this patch moves the code to a helper,  so that it can
be reused by other IOMMU drivers.

Since the resulting function is trivial, it is declared as static
inline in the <linux/mdev.h> header file.

Signed-off-by: Greg Kurz <groug@kaod.org>